wasserstoff/AiInternTask/requirements.txt
torch
torchvision
transformers
pillow
sqlite3
matplotlib
numpy
easyocr
pandas
streamlit